refactor: rename pasteUrl to be more semantically correct
This commit refactors pasteUrl to make it more generic and eliminates
the confusion with the name of the function (paste implies pasting
versus copying). Plus, who knows what else you might want to copy to the
clipboard as part of some future action?

Also, renamed copyUrlToClipboard to be more specific about the current
URL to avoid confusion about copying a URL selected via a mark.